Fix issue with WIPEOUTGRIP + updated REAME

+++ SioForgeCAD/Commun/Mist/Files.cs
@@ -16,7 +16,7 @@ namespace SioForgeCAD.Commun.Mist
                 if (value < 0) { return "-" + SizeSuffix(-value, decimalPlaces); }
 
                 int i = 0;
-                decimal dValue = (decimal)value;
+                decimal dValue = value;
                 while (Math.Round(dValue, decimalPlaces) >= 1000)
                 {
                     dValue /= 1024;
